# testing-environment-backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api-tokens": {
      "get": {
        "operationId": "ApiTokensController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiTokenResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "API tokens"
        ]
      },
      "post": {
        "operationId": "ApiTokensController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateApiTokenResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "API tokens"
        ]
      }
    },
    "/api-tokens/{tokenId}": {
      "delete": {
        "operationId": "ApiTokensController_revoke",
        "parameters": [
          {
            "name": "tokenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiTokenResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "API tokens"
        ]
      }
    },
...
    "/projects/{projectId}/environment-config": {
      "post": {
        "operationId": "EnvironmentConfigsController_create",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertEnvironmentConfigDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Environment configs"
        ]
      },
      "get": {
        "operationId": "EnvironmentConfigsController_find",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Environment configs"
        ]
      },
      "patch": {
        "operationId": "EnvironmentConfigsController_update",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertEnvironmentConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Environment configs"
        ]
      }
    },
    "/projects/{projectId}/environment-config/compile": {
      "post": {
        "operationId": "EnvironmentConfigsController_compile",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompileEnvironmentConfigDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Environment configs"
        ]
      }
    },
    "/projects/{projectId}/environment-config/preflight": {
      "post": {
        "operationId": "EnvironmentConfigsController_runPreflight",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreflightEnvironmentConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvironmentPreflightResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Environment configs"
        ]
      }
    },
    "/projects/{projectId}/environment-config/import-compose": {
      "post": {
        "operationId": "EnvironmentConfigsController_importCompose",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportComposeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComposeImportResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Environment configs"
        ]
      }
    },
    "/projects/{projectId}/environment-config/dry-runs": {
      "post": {
        "operationId": "EnvironmentConfigsController_createDryRun",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEnvironmentDryRunDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvironmentDryRunResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Environment configs"
        ]
      },
      "get": {
        "operationId": "EnvironmentConfigsController_listDryRuns",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedEnvironmentDryRunsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Environment configs"
        ]
      }
    },
    "/projects/{projectId}/environment-config/dry-runs/{dryRunId}": {
      "get": {
        "operationId": "EnvironmentConfigsController_getDryRun",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dryRunId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvironmentDryRunResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Environment configs"
        ]
      }
    },
    "/projects/{projectId}/environment-config/dry-runs/{dryRunId}/cancel": {
      "post": {
        "operationId": "EnvironmentConfigsController_cancelDryRun",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dryRunId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvironmentDryRunResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Environment configs"
        ]
      }
    },
    "/projects/{projectId}/environment-config/revisions": {
      "get": {
        "operationId": "EnvironmentConfigsController_revisions",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Environment configs"
        ]
      }
    },
    "/projects/{projectId}/environment-config/revisions/compare": {
      "get": {
        "operationId": "EnvironmentConfigsController_compareRevisions",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Environment configs"
        ]
      }
    },
    "/projects/{projectId}/environment-config/revisions/{revisionId}": {
      "get": {
        "operationId": "EnvironmentConfigsController_getRevision",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "revisionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvironmentConfigRevisionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Environment configs"
        ]
      }
    },
    "/projects/{projectId}/environment-config/revisions/{revisionId}/publish": {
      "post": {
        "operationId": "EnvironmentConfigsController_publishRevision",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "revisionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Environment configs"
        ]
      }
    },
...
  },
  "info": {
    "title": "Backend Test Runner API",
    "description": "MVP backend control plane and local runner engine.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "ApiTokenResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "projectId": {
            "type": "string",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "revokedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastUsedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "scopes",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateApiTokenDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "projectId": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "scopes"
        ]
      },
      "CreateApiTokenResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "projectId": {
            "type": "string",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "revokedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastUsedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "token": {
            "type": "string",
            "description": "Plaintext token shown once at creation time."
          }
        },
        "required": [
          "id",
          "name",
          "scopes",
          "createdAt",
          "updatedAt",
          "token"
        ]
      },
...
      "UpsertEnvironmentConfigDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "DOCKER_COMPOSE",
              "EXTERNAL_URL"
            ],
            "default": "DOCKER_COMPOSE"
          },
          "composeYaml": {
            "type": "string"
          },
          "backendTestYaml": {
            "type": "string"
          },
          "visualConfig": {
            "type": "object"
          },
          "baseRevisionId": {
            "type": "string",
            "description": "Revision ID the client last loaded; mismatch returns 409 Conflict"
          }
        },
        "required": [
          "type"
        ]
      },
      "CompileEnvironmentConfigDto": {
        "type": "object",
        "properties": {
          "visualConfig": {
            "type": "object"
          }
        },
        "required": [
          "visualConfig"
        ]
      },
      "PreflightEnvironmentConfigDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "DOCKER_COMPOSE",
              "EXTERNAL_URL"
            ]
          },
          "composeYaml": {
            "type": "string"
          },
          "backendTestYaml": {
            "type": "string"
          },
          "visualConfig": {
            "type": "object"
          },
          "revisionId": {
            "type": "string"
          }
        }
      },
...
      "EnvironmentPreflightResultDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "checks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PreflightCheckDto"
            }
          },
          "securityErrors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dependencyWarnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "resourceEstimation": {
            "$ref": "#/components/schemas/EnvironmentResourceEstimationDto"
          }
        },
        "required": [
          "ok",
          "checks",
          "securityErrors",
          "dependencyWarnings",
          "resourceEstimation"
        ]
      },
      "ImportComposeDto": {
        "type": "object",
        "properties": {
          "composeYaml": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "enum": [
              "paste",
              "upload"
            ]
          }
        },
        "required": [
          "composeYaml",
          "source"
        ]
      },
      "ComposeImportResultDto": {
        "type": "object",
        "properties": {
          "visualConfig": {
            "type": "object"
          },
          "analysis": {
            "type": "object"
          },
          "importWarnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "unsupportedFields": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "visualConfig",
          "analysis",
          "importWarnings",
          "unsupportedFields"
        ]
      },
      "CreateEnvironmentDryRunDto": {
        "type": "object",
        "properties": {
          "revisionId": {
            "type": "string"
          }
        },
        "required": [
          "revisionId"
        ]
      },
      "EnvironmentConfigRevisionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "environmentConfigId": {
            "type": "string"
          },
          "revisionNumber": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED"
            ]
          },
          "sourceMode": {
            "type": "string"
          },
          "visualConfig": {
            "type": "object"
          },
          "compiledComposeYaml": {
            "type": "string"
          },
          "compiledRuntimeYaml": {
            "type": "string"
          },
          "schemaVersion": {
            "type": "number"
          },
          "createdById": {
            "type": "string",
            "nullable": true
          },
          "publishedById": {
            "type": "string",
            "nullable": true
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "environmentConfigId",
          "revisionNumber",
          "status",
          "sourceMode",
          "compiledComposeYaml",
          "compiledRuntimeYaml",
          "schemaVersion",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "EnvironmentDryRunResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "environmentConfigRevisionId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "QUEUED",
              "PREPARING_WORKSPACE",
              "VALIDATING_ENVIRONMENT",
              "PULLING_IMAGES",
              "STARTING_ENVIRONMENT",
              "WAITING_FOR_HEALTHCHECK",
              "COLLECTING_LOGS",
              "CLEANING_UP",
              "PASSED",
              "INFRA_FAILED",
              "TIMED_OUT",
              "CANCEL_REQUESTED",
              "CANCELLED"
            ]
          },
          "queueJobId": {
            "type": "string",
            "nullable": true
          },
          "runnerVersion": {
            "type": "string",
            "nullable": true
          },
          "failureCategory": {
            "type": "string",
            "enum": [
              "ENVIRONMENT_VALIDATION",
              "IMAGE_PULL",
              "CONTAINER_START",
              "HEALTHCHECK",
              "NETWORK",
              "TIMEOUT",
              "CANCELLED",
              "INTERNAL"
            ],
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "cancelRequestedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "environmentConfigRevision": {
            "$ref": "#/components/schemas/EnvironmentConfigRevisionResponseDto"
          },
          "logs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EnvironmentDryRunLogResponseDto"
            }
          }
        },
        "required": [
          "id",
          "projectId",
          "environmentConfigRevisionId",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedEnvironmentDryRunsResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EnvironmentDryRunResponseDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
    }
  }
}